Добавляет модули js
